Complete moving build CaseData from AppealCase logic to new class
